requirements.txt
antlr4-python3-runtime==4.9.3
backcall==0.2.0
cycler==0.11.0
dataclasses==0.6
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.4
easydict==1.9
entrypoints==0.4
fonttools==4.27.1
hydra-core==1.2.0
imageio==2.13.0
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.3.1
llvmlite==0.36.0
matplotlib==3.3.4
multiprocess==0.70.12.2
natsort==8.1.0
nest-asyncio==1.5.4
networkx==2.5.1
numba==0.53.1
numpy==1.19.5
odfpy==1.4.1
omegaconf==2.2.1
opencv-python==4.5.4.60
p-tqdm==1.3.3
packaging==21.3
pandas==1.1.5
parso==0.7.1
pathos==0.2.8
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pox==0.3.0
ppft==1.6.6.4
prompt-toolkit==3.0.28
ptyprocess==0.7.0
Pygments==2.11.2
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==22.3.0
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
setuptools-scm==6.3.2
six==1.16.0
threadpoolctl==3.0.0
tifffile==2020.9.3
tomli==1.2.2
torch==1.10.0
torchvision==0.11.1
tornado==6.1
tqdm==4.63.0
traitlets==4.3.3
typing-extensions==3.10.0.2
wcwidth==0.2.5
xgboost==1.2.0
zipp==3.6.0
